factored callback out into a standalone class
