+++ cloud-spanner-r2dbc/src/main/java/com/google/cloud/spanner/r2dbc/v2/DatabaseClientReactiveAdapter.java
@@ -20,8 +20,6 @@ import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutureCallback;
 import com.google.api.core.ApiFutures;
 import com.google.cloud.spanner.AsyncResultSet;
-import com.google.cloud.spanner.AsyncResultSet.CallbackResponse;
-import com.google.cloud.spanner.AsyncResultSet.ReadyCallback;
 import com.google.cloud.spanner.DatabaseAdminClient;
 import com.google.cloud.spanner.DatabaseClient;
 import com.google.cloud.spanner.DatabaseId;
@@ -283,65 +281,7 @@ class DatabaseClientReactiveAdapter {
   private ApiFuture<Void> runSelectStatementAsFlux(
       ReadContext readContext, Statement statement, FluxSink<SpannerClientLibraryRow> sink) {
     AsyncResultSet ars = readContext.executeQueryAsync(statement);
-    return ars.setCallback(REACTOR_EXECUTOR, new ResultSetReadyCallback(sink, ars));
-  }
-
-  static class ResultSetReadyCallback implements ReadyCallback {
-    private FluxSink<SpannerClientLibraryRow> sink;
-
-    private AsyncResultSet spannerResultSet;
-
-    private boolean paused = false;
-
-    ResultSetReadyCallback(FluxSink<SpannerClientLibraryRow> sink, AsyncResultSet resultSet) {
-
-      this.sink = sink;
-      this.spannerResultSet = resultSet;
-
-      this.sink.onRequest(this::unpauseOnAddedDemand);
-      this.sink.onCancel(this.spannerResultSet::cancel);
-      this.sink.onDispose(this.spannerResultSet::close);
-    }
-
-    @Override
-    public CallbackResponse cursorReady(AsyncResultSet resultSet) {
-
-      try {
-        synchronized (this) {
-          if (this.sink.requestedFromDownstream() < 1) {
-
-            // TODO: when googleapis/java-spanner#1192 is released, remove the nested condition,
-            // and return PAUSE regardless of previous state of this.paused. Validate TCK.
-            if (!this.paused) {
-              this.paused = true;
-              return CallbackResponse.PAUSE;
-            }
-          }
-        }
-
-        switch (resultSet.tryNext()) {
-          case DONE:
-            this.sink.complete();
-            return CallbackResponse.DONE;
-          case OK:
-            this.sink.next(new SpannerClientLibraryRow(resultSet.getCurrentRowAsStruct()));
-            return CallbackResponse.CONTINUE;
-          default:
-            // ResultSet returning NOT_READY or null.
-            return CallbackResponse.CONTINUE;
-        }
-      } catch (Throwable t) {
-        this.sink.error(t);
-        return CallbackResponse.DONE;
-      }
-    }
-
-    synchronized void unpauseOnAddedDemand(long request) {
-      if (this.paused && this.sink.requestedFromDownstream() > 0) {
-        this.spannerResultSet.resume();
-        this.paused = false;
-      }
-    }
+    return ars.setCallback(REACTOR_EXECUTOR, new ReactiveResultSetCallback(sink, ars));
   }
 
   private <T> Mono<T> convertFutureToMono(Supplier<ApiFuture<T>> futureSupplier) {
